· Ensure acquired resources are properly destroyed ( not fully working yet, gives a segfault on wl_drm_destroy when destroying the vulkan instance, using vulkan driver built  from latest mesa git sources  )

+++ source/scener/graphics/vulkan/connection.cpp
@@ -68,7 +68,10 @@ namespace scener::graphics::vulkan
     {
         if (_instance)
         {
-            _instance.destroyDebugReportCallbackEXT(_debug_callback, nullptr);
+            if (_debug_callback)
+            {
+                _instance.destroyDebugReportCallbackEXT(_debug_callback, nullptr);                
+            }
             _instance.destroy(nullptr);
         }
     }
@@ -139,18 +142,7 @@ namespace scener::graphics::vulkan
 
     void connection::identify_validation_layers() noexcept
     {
-        std::uint32_t instance_layer_count = 0;
-
-        auto result = vk::enumerateInstanceLayerProperties(&instance_layer_count, nullptr);
-
-        check_result(result);
-
-        Ensures(instance_layer_count > 0);
-
-        std::vector<vk::LayerProperties> instance_layers(instance_layer_count);
-        result = vk::enumerateInstanceLayerProperties(&instance_layer_count, instance_layers.data());
-
-        check_result(result);
+        auto instance_layers = vk::enumerateInstanceLayerProperties();
 
         for (const auto& instance_layer : instance_layers)
         {
@@ -168,69 +160,57 @@ namespace scener::graphics::vulkan
 
     void connection::identify_supported_extensions() noexcept
     {
-        std::uint32_t instance_extension_count = 0;
-
         /* Look for instance extensions */
         vk::Bool32 surfaceExtFound         = VK_FALSE;
         vk::Bool32 platformSurfaceExtFound = VK_FALSE;
         vk::Bool32 debugReportFound        = VK_FALSE;
 
-        auto result = vk::enumerateInstanceExtensionProperties(nullptr, &instance_extension_count, nullptr);
-        check_result(result);
+        auto instance_extensions = vk::enumerateInstanceExtensionProperties(nullptr);
 
-        if (instance_extension_count > 0)
+        for (const auto& instance_extension : instance_extensions)
         {
-            std::vector<vk::ExtensionProperties> instance_extensions(instance_extension_count);
-
-            result = vk::enumerateInstanceExtensionProperties(nullptr, &instance_extension_count, instance_extensions.data());
-
-            check_result(result);
-
-            for (const auto& instance_extension : instance_extensions)
+            if (!strcmp(VK_KHR_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
             {
-                if (!strcmp(VK_KHR_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    surfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_SURFACE_EXTENSION_NAME);
-                }
-                if (!strcmp(VK_EXT_DEBUG_REPORT_NAME, instance_extension.extensionName))
-                {
-                    debugReportFound = 1;
-                    _extension_names.push_back(VK_EXT_DEBUG_REPORT_NAME);
-                }
+                surfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_SURFACE_EXTENSION_NAME);
+            }
+            if (!strcmp(VK_EXT_DEBUG_REPORT_NAME, instance_extension.extensionName))
+            {
+                debugReportFound = 1;
+                _extension_names.push_back(VK_EXT_DEBUG_REPORT_NAME);
+            }
 #if defined(VK_USE_PLATFORM_WIN32_KHR)
-                if (!strcmp(VK_KHR_WIN32_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    platformSurfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_WIN32_SURFACE_EXTENSION_NAME);
-                }
+            if (!strcmp(VK_KHR_WIN32_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
+            {
+                platformSurfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_WIN32_SURFACE_EXTENSION_NAME);
+            }
 #elif defined(VK_USE_PLATFORM_XLIB_KHR)
-                if (!strcmp(VK_KHR_XLIB_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    platformSurfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_XLIB_SURFACE_EXTENSION_NAME);
-                }
+            if (!strcmp(VK_KHR_XLIB_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
+            {
+                platformSurfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_XLIB_SURFACE_EXTENSION_NAME);
+            }
 #elif defined(VK_USE_PLATFORM_XCB_KHR)
-                if (!strcmp(VK_KHR_XCB_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    platformSurfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_XCB_SURFACE_EXTENSION_NAME);
-                }
+            if (!strcmp(VK_KHR_XCB_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
+            {
+                platformSurfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_XCB_SURFACE_EXTENSION_NAME);
+            }
 #elif defined(VK_USE_PLATFORM_WAYLAND_KHR)
-                if (!strcmp(VK_KHR_WAYLAND_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    platformSurfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_WAYLAND_SURFACE_EXTENSION_NAME);
-                }
+            if (!strcmp(VK_KHR_WAYLAND_SURFACE_EXTENSION_NAME, instance_extension.extensionName))
+            {
+                platformSurfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_WAYLAND_SURFACE_EXTENSION_NAME);
+            }
 #elif defined(VK_USE_PLATFORM_MIR_KHR)
 #elif defined(VK_USE_PLATFORM_DISPLAY_KHR)
-                if (!strcmp(VK_KHR_DISPLAY_EXTENSION_NAME, instance_extension.extensionName))
-                {
-                    platformSurfaceExtFound = 1;
-                    _extension_names.push_back(VK_KHR_DISPLAY_EXTENSION_NAME);
-                }
-#endif
+            if (!strcmp(VK_KHR_DISPLAY_EXTENSION_NAME, instance_extension.extensionName))
+            {
+                platformSurfaceExtFound = 1;
+                _extension_names.push_back(VK_KHR_DISPLAY_EXTENSION_NAME);
             }
+#endif
         }
 
         if (!surfaceExtFound)
@@ -251,16 +231,9 @@ namespace scener::graphics::vulkan
 
     void connection::identify_physical_devices() noexcept
     {
-        std::uint32_t gpu_count = 0;
-        auto result = _instance.enumeratePhysicalDevices(&gpu_count, nullptr);
-        check_result(result);
-        Ensures(gpu_count > 0);
+        auto physical_devices = _instance.enumeratePhysicalDevices();
 
-        std::vector<vk::PhysicalDevice> physical_devices(gpu_count);
-        result = _instance.enumeratePhysicalDevices(&gpu_count, physical_devices.data());
-        check_result(result);
-
-        _physical_devices.reserve(gpu_count);
+        _physical_devices.reserve(physical_devices.size());
 
         for (const auto& pd : physical_devices)
         {
